fix issue with element collections with a parent were searching whole dom (#33)
When a parent element is defined, we were properly forcing a leading ".//" so that we would locate child elements in the context of the parent. For element collections, ie createBaseElements(), we were not performing this action. This fix makes sure that only when a parent element is specified, that we anchor the child elements to the parent